Add read .txt file function
Add the function to read the .txt file to the controller.js file.
This function uploads the file to form using multer.
Then, take the path of the generated file and pass it on to the readFile.
And readFile imports some data from the.txt file.
Mark on Read Template.